Migrate anchors into crawler
It doesn't need to know about tiled so it's general
