+++ Packages/LMCore/Runtime/Crawler/IDungeon.cs
@@ -5,7 +5,10 @@ namespace LMCore.Crawler
 {
     public interface IDungeon 
     {
-        public string MapName { get; } 
+        public string MapName { get; }
+        
+        public float GridSize { get; }
+
         public IDungeonNode this[Vector3Int coordinates] { get; }
         public bool HasNodeAt(Vector3Int coordinates);
         public List<IDungeonNode> FindTeleportersById(int id);
