+++ Packages/LMCore/Runtime/Crawler/IMovingCubeFace.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace LMCore.Crawler
+{
+    public interface IMovingCubeFace 
+    {
+        public Vector3 VirtualNodeCenter { get; }
+    }
+}
