+++ Packages/LMCore/Runtime/Crawler/IMovingCubeFace.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 948a5c5e0559c6b43b7f26cc67b94348
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
