+++ Packages/LMCore/Runtime/Crawler/PositionSentinel.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+namespace LMCore.Crawler
+{
+    public class PositionSentinel : MonoBehaviour
+    {
+        [Tooltip("Use None for center of cube face")]
+        public Direction Direction;
+        public Vector3 Position => transform.position;
+    }
+}
