+++ Packages/LMCore/Runtime/Crawler/PositionSentinel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb0cd76f9dbcf6640844f3bdab771246
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
