--- Packages/LMCore/Runtime/TiledDungeonBuilder/PositionSentinel.cs
@@ -1,12 +1,0 @@
-using LMCore.Crawler;
-using UnityEngine;
-
-namespace LMCore.TiledDungeon
-{
-    public class PositionSentinel : MonoBehaviour
-    {
-        [Tooltip("Use None for center of cube face")]
-        public Direction Direction;
-        public Vector3 Position => transform.position;
-    }
-}
