--- Packages/LMCore/Runtime/TiledDungeonBuilder/TDAnchor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 71bb9b26812f15c41b162c99ffcdb766
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
